Refactor to separate out draw utility functions
Certain draw functions do not really belong in the core `screen` code.
This change required an alteration to how the core code is compiled.
Opted for producing a static library at this time.

Given the now multiple headers, it was pertinent to combine them into
a single header for trivial inclusion.